Fix/timetable: fix dayOfWeek serialization, teacher view, and auth issues
- Fix timetable entries showing empty Day column: DayOfWeek was serialized
  as string ("Sunday") by JsonStringEnumConverter but frontend expected int;
  changed TimetableEntryDto.DayOfWeek to int and cast in MapEntryDto
- Add teacher read-only timetable view: teachers see only their own published
  schedule (semester picker + read-only table, no admin actions)
- Fix AcademicController auth: class-level role guard blocked all authenticated
  users from GET endpoints; moved SchoolAdmin/SchoolManager restriction to
  write methods so teachers can load semesters/subjects/grades
- Fix teacher-subject-links: teacher dropdown was binding t.id (undefined)
  instead of t.profileId, causing Guid parse failure on save

+++ src/SkillSphere.Infrastructure/Services/TimetableService.cs
@@ -407,7 +407,7 @@ public class TimetableService : ITimetableService
         TeacherProfileId = e.TeacherProfileId,
         TeacherName = e.TeacherProfile.User.FirstName + " " + e.TeacherProfile.User.LastName,
         RoomId = e.RoomId, RoomName = e.Room.Name, RoomType = e.Room.RoomType,
-        DayOfWeek = e.DayOfWeek,
+        DayOfWeek = (int)e.DayOfWeek,
         PeriodDefinitionId = e.PeriodDefinitionId, PeriodNumber = e.PeriodDefinition.PeriodNumber,
         PeriodLabel = e.PeriodDefinition.Label, StartTime = e.PeriodDefinition.StartTime, EndTime = e.PeriodDefinition.EndTime
     };
